requirements.txt
openai>=1.0.0
pyautogen>=0.2.0
